Move Workplace model to places app
It seems to be the logical place for it. I think Chris suggested an
organizations app on chat, but so far the places app is quite small and
feels like a decent fit.
